Handle (CharAt() & 65535) case
This case happens when casting the result of CharAt() to integer and then cast the integer back to char.
For this case, the bitwise and is removed when modeling to Z3 SMT formula since it is unnecessary.

For a more general case (bitwise AND between two arbitrary integers), both integers are converted to bitvectors, AND with bvand and the result is converted back to integer.
TODO: this approach is impractically slow, find a more efficient workaround for integer AND modeling.